+++ app/src/main/java/com/dms/pmsandroid/domain/calendar/entity/EventModel.kt
@@ -1,0 +1,3 @@
+package com.dms.pmsandroid.domain.calendar.entity
+
+data class EventModel(val eventNames: List<String>, val eventTypes: List<EventTypes>)
