+++ app/src/main/java/com/dms/pmsandroid/domain/calendar/entity/EventTypes.kt
@@ -1,0 +1,5 @@
+package com.dms.pmsandroid.domain.calendar.entity
+
+enum class EventTypes {
+    MUST_GO_HOME, EXAM, VACATION, HOLIDAYS, ETC
+}
