+++ app/src/main/java/com/dms/pmsandroid/domain/calendar/repository/CalendarRepository.kt
@@ -1,6 +1,6 @@
 package com.dms.pmsandroid.domain.calendar.repository
 
-import com.dms.pmsandroid.presentation.feature.calendar.model.EventModel
+import com.dms.pmsandroid.domain.calendar.entity.EventModel
 import io.reactivex.rxjava3.core.Single
 import java.time.LocalDate
