+++ app/src/main/java/com/dms/pmsandroid/domain/meal/entity/Meal.kt
@@ -1,0 +1,3 @@
+package com.dms.pmsandroid.domain.meal.entity
+
+data class Meal(val breakfast: ArrayList<String>?, val lunch: ArrayList<String>?, val dinner: ArrayList<String>?)
