+++ app/src/main/java/com/dms/pmsandroid/presentation/di/module/CalendarModule.kt
@@ -1,6 +1,6 @@
 package com.dms.pmsandroid.presentation.di.module
 
-import com.dms.pmsandroid.data.calendar.remote.ProvideCalendarApi
+import com.dms.pmsandroid.data.calendar.remote.CalendarRemoteDatasource
 import com.dms.pmsandroid.domain.calendar.repository.CalendarRepository
 import com.dms.pmsandroid.data.calendar.repository.CalendarRepositoryImpl
 import com.dms.pmsandroid.presentation.feature.calendar.viewmodel.CalendarViewModel
@@ -10,7 +10,7 @@ import org.koin.dsl.module
 val calendarModule = module {
     single<CalendarRepository> { CalendarRepositoryImpl(get(), get()) }
 
-    single { ProvideCalendarApi(get(), get()) }
+    single { CalendarRemoteDatasource(get(), get()) }
 
     viewModel { CalendarViewModel(get()) }
 }
