--- app/src/main/java/com/dms/pmsandroid/presentation/feature/calendar/model/EventKeyModel.kt
@@ -1,3 +1,0 @@
-package com.dms.pmsandroid.presentation.feature.calendar.model
-
-data class EventKeyModel(val month: Int, val day: String)
