--- app/src/main/java/com/dms/pmsandroid/presentation/feature/calendar/model/EventModel.kt
@@ -1,5 +1,0 @@
-package com.dms.pmsandroid.presentation.feature.calendar.model
-
-import com.dms.pmsandroid.data.calendar.EventTypes
-
-data class EventModel(val eventNames: List<String>, val eventTypes: List<EventTypes>)
